+++ src/icons.h
@@ -8,6 +8,7 @@ extern const uint8_t PROGMEM iconOn[];
 extern const uint8_t PROGMEM iconOff[];
 extern const uint8_t PROGMEM btIcon[];
 extern const uint8_t PROGMEM usbIcon[];
+extern const uint8_t PROGMEM chargingIcon[];
 extern const uint8_t PROGMEM splashBitmap[];
 extern const uint8_t PROGMEM splashVolumeBitmap[];
 extern const uint8_t PROGMEM splashGameBitmap[];